Implement & test deep copy
Not sure whether to call it deep copy or deep clone..

Anyways, there should (instead of this?) be a helper with which you can
clone and change the parent at the same time. Just assigning the Child
somewhere else won't change the parent id. Should be easy to implement
by placing the parent ID on the deep_copy_id_stack.